Disable libdeflate use for libtiff
As part of the Ghostscript/GhostPDL configure script, we also
run the libtiff configure. If libdeflate was present, and
running Ubuntu 21.04 (or later?) and using our shipped
libtiff sources, it would attempt to build with libdeflate
support, but not include the library to link to, and
cause a build error.

We don't need libdeflate support, so disable it explicitly
when we call the libtiff configure

Also, add an environment variable (GS_TIFF_CONFIGURE_OPTS) to
the libtiff configure command line, as an easy "end user"
workaround for similar future problems. So doing:

./configure GS_TIFF_CONFIGURE_OPTS="--disable-libdeflate"

Would work around the libdeflate issue.