Fix Bug 703898. Type 5 Halftone ignored colorants that were not defined.
The customer's PDF file defined a Type 5 halftone with different halftones
for some of the named components, but only Default and Cyan, Magenta, Yellow,
and Black were actually being generated and installed into the pgs->dev_ht
so the /Default would be used for all the spot colors.

This change adds the colorant that is not yet known is when it is seen by
gs_color_name_component_number rather than when the colorspace is set with
the colorant. The equivalent color is still loaded when the colorspace is
loaded.

Note that for devices that support DeviceN (extra colorants), we do not
add the "Red", "Green", "Blue" and "Gray" when seen in the HalftoneType,
but only AFTER they are set in a colorspace. This is needed since many
files set these other component names even when they are not yet entered
as spot colors (in Separation or DeviceN colorspaces).