Be more tolerant of of broken CFF streams.
In this case (Bug696716.pdf) the stream has a few bytes (probably corrupt) at
the end of the stream, which look like that start of a CFF encoding integer, but
not enough bytes to make up an integer. So we run out of data, and error.

Spot that we are at the end of the stream, and don't error out.

A stream broken in an earlier, more critical way will error out later due to
lack of certain vital things like CharStrings, Subrs, etc...