Don't overwrite existing CharString entries
For the purpose of the CharStrings dict in a Type 1 font in a PDF, "def"
shouldn't overwrite a key that already exists - i.e. later definitions do not
take precidence over earlier ones, see:
https://bugs.ghostscript.com/show_bug.cgi?id=703454

This meant rejigging where we add a "fake" .notdef glyph to cope with bad fonts
which don't define one. Originally this was done when we create the charstrings
dict, it's now done later when we patch up for "special" glyph names - it will
add a .notdef if one is not present.

Since the charstring "patching" shouldn't be necessary, it doesn't return an
error, so just in case the adding of the .notdef failed, we check for it
being missing before it in the FAPI code (it used to assume it was there).

Lastly, as part of changing the FAPI code, (at the suggestion of another
member of the team) remove the overloading of the "code" variable, so it
only comes into effect for error codes.