For vertical fonts apply Tc and Tw to the y axis, not x
We were always applying the Tc (character spacing) and Tw (word spacing)
parameters to the x value, for vertical writing fonts we need to apply
them to the y value instead.

Fixes a few test files completely, improves others. A few seem worse but
I believe this is a different (possibly font-related) problem and this
fix merely makes the problem more obvious.