Add -dRESETRESOURCES option for gpdl.
In this instance 'RESOURCES' means fonts, macros, symbol sets etc.
This applies to PCL (and maybe others), but not Postscript
resources.

Default (and setting this to 0) gives the behaviour where resources
are never reset based upon job language changes.

Setting this to 1 causes resources to be reset between jobs (whenever
we return to PJL).

Setting this to 2 causes resources to be reset whenever we "change
language". For this setting, we do not count returning to PJL and
reentering the same language as being a language change.

(Includes fixes and code from Henry Stiles).