Fixes for Annotations with no AP
Sample for all of these was: tests_private/pdf/uploads/annots-noAP.pdf

1) Fix math to deal with division by 0 in /Ink

2) Don't substitute Draft for unknown Annotation Stamp Names

AR has a bunch of built-in stamps that we don't support if no AP
is provided.

Note that on page 1 of this file we render the large FreeText annotation
in white, so it doesn't show up. This is because the /DA sets the color
to white. In gs, since there is an error setting the "Arial" font, it punts
the whole DA string.  In pdfi we set the color and font (which picks up
Helvetica) with no error.  So we pdfi uses white, like it says, and gs
uses black after getting an error.  AR uses black, but it is most likely
ignoring the DA and picking up the xml and DS, which says the color is black.