Set 'single_byte_space' for text so that we only apply Tw to real spaces
The Tw parameter is only supposed to be applied to characters with the
single byte code 32, multi-byte character codes with the value 32 should
not be additionally spaced.

Because of the different requirements of PDF and PostScript, we set a
flag in the enumerator if we only want to apply word spacing to single
byte chracters.