Check cmap (3,1) table for AGL unicode value
If we find a match for the glyph name in the AGL, we're supposed to use the
Unicode code point that matches it. But it turns out we also need to check
whether the codepoint exists in the 3,1 cmap table, and if not, fall back
to trying to use the font's post table.