Change get_color_mapping_procs device procedure.
Previously, devices would call 'get_color_mapping_procs' to get
a procedure table from a device. They would then call those
procs, passing in the device as an argument.

For forwarding/subclassing devices, this is problematic, because
getting a child device's color_mapping_procs, and passing it a
parent's device pointer will lead to failures. Accordingly lots of
wrapping functions have to be written.

Here we try an alternative formulation. When we
'get_color_mapping_procs' we get both the procedure table, and
the device pointer to pass to them. For most devices this will
be the incoming device pointer parrotted back. For forwarding
and subclass devices, it will be the appropriate child device.