Rejig font loading for Tj
Previously pdfi_Tj using the SubType key in the font object dictionary to decide
to which font font type handler it should dispatch the font loading.

This in inadequate on two fronts: the font SubType key is not sufficient to
accurately determine the font type (nor is FontFile/FontFile2/FontFile3 nor
the FontFile3 SubType!). Secondly, it hampers handling of non-embedded fonts
because in the non-embedded case, the font we end up using may not match the
font SubType key at all.

The first problem meant each font loader function was checking the stream magic
number, and pushing off to a different handler if required - which is less
than ideal!

So, centralise that process: introduce generic pdfi_load_resource_font() and
pdfi_load_font() functions which take care of all the font type identifying,
and also deal with the non-embedded situation. Thus the invididual font handlers
don't have to.