Make bmpcmp adapt to different spot colorant order in psdcmyk files.
When processing the second image file, if the CMYK equivalent color for a
spot color does not match that of the first file check if there was a
colorant with the same CMYK in the first file, and if found, map the
color in the second file to the position used in the first file.

As the second file is stored in the page memory, the color_map is used
to store the colorant in the same position as was used for the first
image file.

Also get rid of the globals and change the psd_read to use (void) for
reading the data when it is going to discard the result (for clarity).
(eliminating the globals for the spots is from Robin Watts, extended
to allow for the addition of a color_map).