Fix valgrind problem with very small files
The code scanning backwards in the file for the 'startxref' copies
leftover data from the start of the buffer to the end, if it reads
more than one buffer of data. It also reduces the number of bytes
to be read so that the next buffer of data ends up exactly joining the
preserved data.

There were two problems however; firstly if the file was very small so
that a single 2Kb buffer contained the whole file, then the calculation
of the point to start scanning was incorrect, leading to us reading
from uninitialised memory.

Secondly, if we failed to read as many bytes as required to fill the
buffer up to the point of a preserved buffer, then we would leave a 'gap'
between. This is unlikely to be a real problem, since we would have
read back to the beginning of the file, but for very small files it is
just possible this could happen.