Fix PCL Monochrome Palette device to work with pdfwritten text.
The pdfwrite device, when called to begin_text, looks at the
color passed to it, and uses the 'high level' version of the
color (i.e. the non monochromised version).

We can avoid this by marking the ccolor_valid flags to be 0 as
the call passes through the pcl_mono device.

The implementation of this is a bit grotty.