Rejig text_begin dev_proc.
In this commit we remove several parameters to the text_begin
device procedure, because they are implied.

Firstly, we remove pdcolor, and just use the device color from
the current color in the graphics state (pgs->color[0]). This is
actually what the code did anyway (in that it was assumed that
pdcolor, if it was non-NULL was a pointer to this).

Similarly, we remove the path param, as (if used) this is always
the same as pgs->path.

Finally, we remove the memory param as this is (almost always)
the same as pgs->memory (and in cases where it isn't, tests indicate
that it could be).

This has knock on effects for the text enumerator structure.