Headers required to fix bmpcmp builds in the cluster.
In order to build a bmpcmp binary, we can no longer rely on the tif
configuration headers built as part of a ghostscript build. This
is possibly down to GPDL supporting more variants of tif.

So to make life easier, we keep 2 tiff config headers here that
the cluster can unpack as part of the process of building bmpcmp.