Add psdcmyktags and psdcmyktags16 devices
Add example tags devn devices, both 8 and 16 bit. The tags
plane is always placed as the last plane in the data (i.e
it follows any spots that get added). Changes were made to
remove tags conditionals in the planar memory device code,
since it really should not care what the extra planes are
used for and it should not add components to the device based
upon the tags support. The target device should handle any
of this sort of setup. There were
also some changes needed in the pdf14 code, as the tags
information was not getting properly communicated when we
had knockout objects and devn colors.

Also, fix various whitespace issues.