Implement heuristic glyph width "matching".
Where a mismatched substitute font is loaded from file, issues arise when the
metrics in the Widths array don't match those native to the font - especially
when those in the Widths array are smaller than the native ones.

This compares the metrics for (we hope!) the A-Z,a-z glyphs, and scales the font
down (only down) if the glyph metrics advance width is greater than the Widths
value.

This implementation uses the "glyph_info" font proc to retrieve the metrics,
compared to the Ghostscript implementation which uses stringwidth - thus
there is a a considerable of scaling calculations that this implementation
doesn't do to arrive at the desired values. So there may be (hopefully small)
mismathes with the Ghostscript output.