Have pdfi use the current PostScript graphics state on each page
Rather than copying the graphics state, have pdfi use the PostScript
graphics state. This allows us to alter the graphics state between
pages, for example to do N-up organisation.

There are a few wrinkles to this.

Firstly ICC profiles have to be created using the PostScript memory
allocator, but also in stable memory. This is because they are stored in
the profile cache which can persist until the interpreter shuts down.
We can't have them disappearing due to a restore.

Similarly for patterns, but in this case because of patterns involving
transparency. These can end up creating a patt_trans structure, and that
needs to be allocated in the graphics state memory allocator.

Then there's the 'interpreter part of the graphics state'. This
absolutely must be correct for whichever interpreter is currently
running. If pdfi then it must be a pdfi interpreter context, if GS
then it must be a PostScript interpreter context. There's a bunch of
things stored in here, but we just treat them opaquely.

The odd thing (to me) is that the interpreter context does not seem to
obey gsave/grestore. When we grestore we copy the interpreter state from
the 'current' gstate to the 'saved' gstate. This seems odd to me, but
it means we cannot use grestore to just go back to the point before
we made a PS gstate into a pdfi gstate, we have to be more careful
than that.