Pattern instance life span problem with pdfi in gs
When pdfi is called from gs (i.e. using the gs graphics state), the pattern
tile cache exists outside, and lasts beyond the "life", of the pdfi instance.

But pattern instances do not. That causes a crash when Ghostscript runs the
end of job restore, and the garbager tries to access the pattern instance, via
the clist accumulator device in the pattern tile cache.

So, have pdfi remove tile cache entries for relevant patterns when the pattern
instances get freed.