Prevent a seg fault on unrecoverable errors when running in PostScript.
If the PDF file failed to open properly because it is invalid and
can't be repaired, we need to throw away the 'PDF copy' of the input
stream. Otherwise we get a seg fault trying to close it on exit when it
has already been closed by the PostScript interpreter.