Fix 'off by one' error in transfer function sampling.
When creating the 'map' for transfer functions, as well as black
generation and under colour removal, we divided by the transfer map
size, when it should have been the map size -1. This was leading to very
slightly too small sample values (differing in the 4th decimal place)
being fed to the transfer function for sampling.

Mostly this was too small to show any problems, but a few files have
'special' transfer functions and sampling those incorrectly led to
visible artefacts.