Address compiler warning of bug 703887.
As Alex Cherepanov points out, this can be addressed by changing the
calling sequence as was done in commit 36e09729a by Ken for bug 703886.

Note that this bug mentions the warning resuted from gcc using the
option -D_FORTIFY_SOURCE=2.

Note that the function used by these routines dict_proc_array_param,
verifies that the array size is correct and issues a 'rangecheck' if
not, so making this a generic ref * when it requires an array of 3
refs is not a problem.