Dereference DeviceN process space when finalizing a DeviceN space
The finalize routine for the DeviceN space cleans up all the allocated
objects stored in the DeviceN space, including names, colorants and so
forth.

However, it doesn't count down the reference to the process space of
the DeviceN space. In Ghostscript the garbage collector will eventually
free this, but in pdfi it leads to a memory leak.