Fix string source streams' ownership of the string source
Until relatively recently there was confusion in the graphics library
about the ownership (and therefore lifetime) of strings uses as the
source to a stream, such as for ReusableStreams.

This was fixed in bd56e2ddd62e5706dfafdfe817d71e6d27a7fe49 and at the
time we thought that pdfi didn't suffer from the problem. However it
turns out that functions and shadings do have this problem. There is
no time (for these objects) when we can safely free the string in the
interpreter, we need the stream to do that when it is closed. For other
objects we want to manage the lifetime ourselves.

So add a new parameter to the pdfi_open_memory_stream_* functions to
determine whether the ownership of the string is transferred to the
stream or remains with the caller.