pdfwrite: put back saved text enumerator->text data
When scanning CMap decoded strings (scan_cmap_text), we accumulate a list of
gs_glyph values from the original string/byte buffer passed using the text
enumerator.

Once we have that, we replace the text/string buffer in the enumerator with
that accumulated list of gs_glyph values, in order to pass that information
for further processing in pdfwrite.

The problem was, we never put the original buffer pointer back in place (as well
as things like the current index into the buffer). Since that buffer is also a
temporary one allocated a level above (process_cmap_text) where it also gets
swapped in and out, the original, interpreter allocated string/byte buffer
remains intact, and in place when we exit.

The problem arises, depending on the memory manager in play at the time, when
process_cmap_text() tries to free the buffer it allocated, because the pointer
in the enumerator is no longer the correct one.

The solution is to simply put back the buffer, after we've finished with the
accumulated list of gs_glyph values.