Fix the string matching for font substitution
Originally, the string matching took a pdfi name object, and used the length
of the name object to limit the amount string entries it compared.

This doesn't work reliably because the string comparison intentionally ignores
spaces in either string, thus we could ignore spaces, and read off the end
of the string object data.

Since we already allocated a sizeable stack buffer for the file name, reuse that
to hold a null terminated copy of the font name, meaning both strings are
guaranteed null terminated, and we can rely on that to limit the comparison.

If the font name is larger than the file name we can store, you're stuffed
anyway!