Fix SEGVs seen in alldevs tests.
We see SEGVs in the alldevs tests with the escp device where
map_color_rgb is being called on an 8-bit memory device. The
implementation assumes that there will be a palette record
and so indexes into it.

It seems that palettes are actually unusual; the only time we
get one is if the device was created using the makewordimagedevice
operator from postscript.

Accordingly, we here update the map_color_rgb routine to behave
differently in the no-palette-present case. We return rgb values
as if the color index supplied was a greyscale value.

I believe this fixes the same issue seen with some other devices
including lj4dith and lj4dithp.