Fix SEGVs seen in stc device with alldevs test.
It seems that since 2002 at least, the stc device has been
potentially stack smashing.

Back in 2002, the color procedures of the device were updated.
As a result, when in the default cmyk (or cmyk10) modes, the
map_color_rgb routine was set to stc_map_color_cmyk (or
stc_map_color_cmyk10). These write 4 colour components into
an array that is only expecting 3, potentially causing crashes.

Here we revise the naming of the functions to be clearer, and
explicitly write decode/encode_color routines as required.

This should be sufficient to fix the SEGVs. It's hard to verify that
these routines are now doing the correct thing, as we have no
way of testing the output bitstream. If someone is still using
the devices, we trust they'll contact us if it's broken.