Rework error and warning recording and reporting
It has become clear we need more than 32 unique errors and 32
warnings. Rather than choose a larger fixed object to hold the flags
we've decided to go with a more flexible (if more complex) system and
hold the error/warning flags in an array of bytes, but have each error
and warning have its own integer number rather than a flag.

This involves a little arithmetic to turn the error number into a byte
index into the array of flags and a bit within the byte.

Since this involves considerable rework, take advantage of this
opportunity to add more flexibility to the error and warning code.

We now have inline functions to record errors and warnings. If verbose
warnings and errors are selected (which they are by default for debug
builds) then every time we set an error or warning we will write
information to stderr or stdout.

We can now opt to have any or all of; a Ghostscript error code, the
graphics library routine which returned that code, the pdfi error
number, the pdfi routine name, and an extra string full of additional
information. There's a similar routine for warnings.

If we supply a GS error code then the error/warning routines will
turn it into a PostScript error string; if we supply a pdfi error or
warning number then that will be used to select an additional
(hopefully) meaningful text string. extra information is printed
verbatim on a new line.

Currently I anticipate that we will normally use the gs error code along
with the library routine name, but there's no reason why we couldn't
use the gs error code with the pdfi routine name to indicate that
we are returning that error from the pdfi routine.

I've also added a pdfi_log_info function, though that's not widely
used yet. I plan to turn at least some of the existing dbgprintf
usage into pdfi_info() calls in future.

All of these routines respect -dQUIET (or -q) as well as either
-dVerboseErrors or -dVerboseWarnings. -dQUIET trumps verbosity.