Take more care with CFF endchar parameters
The pseudo-SEAC in CFF was getting a param of -28 (then converted to unsigned),
so the lookup in  the built-in StandardEncoding was going badly wrong, and
causing more bad things - segfault, valgrind issue, etc.

This just adds the needed checking. If the value(s) are nonsensical, drop
back to a .notdef glyph. It's not clear what other consumers do, but since they
don't error, .notdef seems a decent first guess.