Fix memory 'leak' with ps2write
When converting a shading into an image ps2write makes a copy of the
current graphics state. If it then decides that it doesn't need to
bother with the shading (zero area fill) it then exits but, crucially,
does not free the copied graphics state.

In PostScript the garbage collector eventually cleans this up but for
pdfi it leaks, and worse than that, under pdfi+GS, when the GC does
clean it up, the pdfi memory has gone and crashes abound.

Free the gstate if we exit the path early.