Error out with broken CFF FDArray font
Originally, the code tried to carry on when there were invalid entries in the
FDArray on the basis that, if those invalid entries were never used, we could
feasibly render the file.

Turns out, the graphics library requires the FDArray to be fully and correctly
populated, so that cannot work.

We now error out, and fall back to a substitute font.