Fix transparent patterns with nullpage
There is a problem with transparent patterns, filling glyphs, when the
device does not support transparency *at all* and the glyph is
sufficiently small to be cached.

Filling a glyph bitmap with a transparent pattern does not, apparently,
work correctly, so the glyph cache is disabled for this case, and there
is code in FAPI which attempts to detect whether a glyph is about to be
filled with a transparent pattern. If it is FAPI will produce a glyph
path, which the graphics library will use to clip the pattern.

If the device does not support transparency at all, this code, quite
reasonably, uses a cached bitmap instead.

Meanwhile the code for filling a glyph with a transparent pattern did not
actually check whether any device was capable of handling the
transparent pattern. It simply assumed that it was, and proceeded to
attempt to fill it.

In the pattern code this leads to us filling in the 'tmask' member of
the gx_color_tile structure, because we are getting a type 1 mask
bitmap from the character cache.

But the pattern painting code would attempt to use the 'tbits' member
of the gx_color_tile because it thinks its filling a path. This leads to
us using uninitialised memory with unpredictable results. Sometimes an
infinite loop, sometimes a seg fault, sometimes incorrect rendering.

This doesn't show up on our testing because we don't generally use the
nullpage device, and other devices either push the PDF14 compositor,
or (pdfwrite) handle all the transparency themselves. I am using the
nullpage device for testing memory leaks on the new PDF interpreter
where this exhibited as a hang. Testing on Ghostscript with the same
device showed the same problem.

In order for this to work correctly FAPI and the pattern painting code
need to agree on whether to use a path or a mask. I tried simply using
the existing is_pdf14_device sepc_op but that doesn't work properly
with pdfwrite, because that is not the pdf14 compositor but does handle
transparency. Making pdfwrite return true for is_pdf14_device would
cause other problems, as that spec_op is widely used and often in cases
that will only work properly if the device really is the PDF14
compositor.

So I have reluctantly added yet another spec_op to indicate that a
device is capable of handling transparent patterns. The PDF14 compositor
and the pdfwrite device both indicate they can handle this, all other
devices indicate they cannot.

FAPI now uses this (instead of testing device methods) to determine
whether it should use a mask bitmap or an outline path. The pattern
instantiation in gs_pattern1_make_pattern() no longer simply uses the
request in the pattern template but checks to see whether the device
permits the use of transparent patterns.