Fix handling of collections that are not type PDF
If there were multiple things in a collection and the last one didn't
happen to be a PDF file, it was getting to the bottom with a code=-1,
which meant it didn't do the processing at the end correctly and
ended up crashing because pointers that were supposed to be setup were
NULL.

Also changed the type of FileNames to a pdf_array since it is one.

Sample is file visual_comm_cs3_examprep.pdf from bugzilla 698946.
Should we add this to the cluster?