Fix memory leak with loop detector and NULL objects
When dereferencing an object, under some failure conditions we can
end up allocating a null object. If we were checking for circular
references we would then add that object number (0) to the loop detector
array of objects encountered.

But 0 isn't a valid object number, and we use it to make a 'mark' in
the array. Adding extra marks really defeats the cleartomark function
of the loop detector. This potentially means we would incorrectly
detect circular references, and also that (because the array was not
empty in cleartomark) we would fail to free the array.

Added code so that we don't add null objects to the loop detector array,
also added code to free the array on exit if it is still present, and
emit a warning in debug builds. Added a comment that we could probably
get a performance improvement by not continually freeing the array. But
for  now it's a useful way to detect this kind of problem.