Fix memory leak with Form XObjects with /Contents
This is, of course, technically invalid. Form XObjects ought to be
streams, and should not contain a /Contents array.

The code to deal with this needs to create a stream object for the
XObject by using the dictionary and the /Contents member of that
dictionary. We need this for the existing code to work.

To do this we were merging the newly created stream dictionary with the
existing form dictionary, but that meant that we copied the /Contents
member as well. This results in the newly created stream object
containing a /Contents array which references the stream object. This
means the object is effectively self-referencing and therefore it's
reference count never reaches zero, leading to a memory leak.

We don't need the /Contents member of the form dictionary, so just
explicitly remove it after the merge.