Don't go back past enclosing stream 'save' when finding mark on stack
We keep a note of the stack index at the point we start interpreting
a stream, and we won't allow 'pop' to remove objects before that. But
we were allowing pdfi_count_to_mark() to go back past that point
looking for a mark on the stack.

This could lead to memory leaks because array_from_stack() would count
back past the 'save' point looking for a mark. Assuming it found one
it would then load that many objects into the array (without checking
the save point) and pop that many objects from the stack.

Because pop() didn't remove the object from the stack, this could lead
to the object having too high a reference count leading to leaks.

This is a rare situation; a file would have to leave entries on the
stack which include an opening mark, but not close it, then start a
content stream, that stream would then have to put entries on the stack
up to a closing mark, without a previous opening mark. The closing mark
would allocate and populate the object leading to the memory leak.