Signal an error for invalid compressed objects
While processing the offsets for the objects in a compressed object
stream if we encountered a non-integer object we would exit (reasonably
since the offsets must be integers) but we didn't actually raise an
error.

This led to us trying to process a NULL object which caused a seg fault.

Fix this by returning an error; for a belt and braces approach make sure
the returned object is not NULL as well, before processing it.