Don't try to use AcroForm for resources if there is none.
When processing a DA entry for a Ch or Tx annotation we search for
Resources in the AcroForm entry. However we have a test file which
has a Ch with a /DA but does not have a /AcroForm.

Trying to use a NULL dictionary for Resources causes seg faults, so
if AcroForm doesn't exist, don't use it. Instead carry on using the
annotation dictionary for Resources, if the Resource isn't defined it
(obviously) can't be used and may cause an error, but it won't crash.