Prevent buffer underrun on CCITTFaxDecode
With the file \tests_private\pdf\customer394\problems\normal_994.pdf
the CCITTFaxDecode filter runs off the beginning of the 'lbuf'
buffer at the end of the image data, in invert_data().

This appears to be 'something' to do with returning to the interpreter
for more data, because that seems to decrement the pointer to one byte
before the beginning of the buffer. This may be a red herring.

What is certainly true is that we actually do write one byte of data
to the byte immediately preceding the start of the buffer. The way
the memory manager currently works means that corrupting this byte
does not cause Ghostscript to see any problems. However when running
pdfi with leak detection it writes to the 'padding' of the memory
structure, corrupting it, and causing an error when the buffer is
freed.

We obviously should not be writing to a point before the beginning of
the buffer, but I cannot make head or tail of this code. It is very
complex (needlessly so IMO), makes extensive use of macros, and of
course is wrapped up with the Ghostscript stream code. The result is
all but incomprehensible, even to the original author I suspect!

Throwing an error when we try to write to a point before the beginning
of the buffer causes a number of files to exit with errors that
previously did not, so its clear that there is something up here, and
exiting with an error is not a fix.

So given that I can't understand the code, and a naive fix is
insufficient, this commit works around the problem by allocating slop
at the beginning of the buffer, to match the slop already allocated at
the end of the buffer.

This is complicated by the fact that the code swaps the two buffers
'lbuf' and 'lprev' as it runs, and by the garbage collector which may
relocate the buffers during the decoding.

So in addition to allocating more memory, we need to recalculate the
lpuf and lprev pointers on entry to the process routines, and when the
lbuf and lprev pointers are swapped we need to swap lprevstart and
lbufstart as well.