Handle Root /Pages entry being a Page dictionary
The Pages entry in the Root catalog is supposed to be an indirect
pointer to the root of the document's page tree. We have examples of
files where it in fact points directly to the sole Page dictionary
in the file.

Acrobat can correct this, as can Ghostscript, so we need to be able to
as well, even though it's invalid.

When opening the document; if the Pages dictionary does not contain a
/Count key we check to see if it has a Type and it's Type is /Page
(rather than /Pages). If so assume there is only one page.

When processing the single page, if PagesTree is NULL assume we have a
single /Page referenced from the Root /Pages entry and retrieve that
instead of the usual search for the page in the page tree. We do check
to ensure that the retrieved dictionary is a Page dictionary.