Fix bug 703996. BT/ET not preserving path around text.
The customer's file is missing circles around numbers. This is because
the file defined a path to draw the circle, then did a BT .. ET to
output the number, followed by 'S' to stroke the path (circle).

The internals of the PDF 'Show' would do a 'newpath' which would clear
the path, so the 'S' stroke would have an empty path.

The fix is to save the path (using 'upath') around the BT .. ET then
restore it afterwards.

Rather than doing { currentpoint } stopped we add a .currentpoint_valid
operator and use that since testing with the PDF 1.7 reference manual
showed that the stopped added 4% to the execution. Other places where
this was done now use this new operator and it speeds up execution by
a bit on that same file (only 1%), but at least it's no slower.