Detect circular reference in Form XObject-based SMask
An SMask which uses a Form XObject can include references to itself
(eg the Form XObject has an SMask entry whose value is the same Form).

We can detect this by installing the loop detector when we process an
smask. This is slightly complicated by the fact that the SMask can
also be a direct object rather than an indirect reference, and in that
case we must not (and don't need to) check it.