Fix reference counting problem with broken file
It is possible on accessing the trailer dictionary in files where the
xref table is broken, for us to need to repair the file. The action of
repairing the file can replace the Trailer dictionary entry in the
context. This obviously counts down the reference and potentially
frees the dictionary.

If we are in the middle of using it, bad things happen.

So take a specific reference in the case where that might happen.