Move PDFSTOPONERROR checks out of the bottom level code
Originally I was handling this at the bottom level of the interpreter,
the interface to the graphics library, but that changed during
development. For various good reasons we don't generally (things like
executing Form streams are an exception) want to detect this at any
point lower than interpreting a content stream.

This allows us to faithfully flag up errors, but still continue with the
remainder of the stream/PDF file if requested.

Many places were swallowing the error at a low level which meant we
were not reporting it. This commit removes that, pushing the error up a
level where it gets reported, while still allowing us to stop if the
flag is set.