General tidying up of loop detection usage
The pdfi_loop_detector_mark() is intended to set a point where we can
clear the stack of objects to be checked for recursion. We need to check
the stack in various places, but we need to be able to clear it back
when we exit certain levels.

For example when we're checking Colour Spaces, when we've checked one
space we need to clear the stack of any objects that were only used in
that colour space. That allows us to check the next colour space, which
might legally use some of the same objects, without erroneously
detecting a circular reference.

It is important that we clear the stack back to that mark when we are
done with a particular loop check. If we don't we can incorrectly
flag circular references.

Some places weren't clearing the stack under some conditions, and some
places weren't checking to make sure the mark/check/clear sequence
succeeded.