Prevent memory leak on image error
If gx_begin_image4 encountered an error from gx_imege_enum_begin, such
as an image Matrix of [0 0 0 0 0 0] then the error case neither
transferred the allocated enumerator to the caller, nor freed it.

This led to a memory leak, as usual PostScript is saved by the garbage
collector. I'm not sure if other languages can use type 4 images, but
pdfi does and the leak showed up there.